feat: Java 17 compatibility for async-indexing module and move index classes
Java 17 compatibility:
- Removed reflection-based access to JDK internals in AsyncUtils
- Replaced with ThreadLocal-based async context management
- Updated AsyncTaskManager to pass context via constructor
- Modified AsyncAspect to handle context properly
- Updated tests to manually set context for Java 17 compatibility
- All tests now pass without --add-opens JVM flags

Refactoring:
- Moved IndexConfigProvider, IndexException, IndexUtils, and ResourceIndex
  from service module to async-indexing module
- Service dependency retained due to JsonConversionFactory and
  GenericElastic7Configuration dependencies

This resolves reflection errors when running with Java 17's stronger
encapsulation of JDK internals.
